Backend/requirements.txt
Django==4.1.3
django-environ==0.9.0
djangorestframework==3.14.0
django-cors-headers==3.14.0
